requirements.txt
selenium==4.10.0
PyPDF2==3.0.1
transformers==4.34.0
prettytable==3.7.0
PyGithub==1.59.0
